Move command line parsing functions into app/embed.cpp
