+++ src/audio/EmbedAudio.cpp
@@ -27,43 +27,6 @@ namespace fs = std::filesystem;
 
 static bool bestQuality = true;
 
-void showUsage(std::string programName) { 
-  fmt::fprintf(cerr, 
-      "Usage: %s [audio_file] [image_file]\n %s %s %s", programName, 
-      "Options:\n", 
-      "\t-h, --help\t\tShow this help message\n", 
-      "\t-f, --fast\t\tLower image quality\n");
-} 
-
-bool meetsReq(int argc, char** argv) {
-  if (argc <= 1 || !(argc > 1 && argc < 3)) {
-    showUsage(argv[0]);
-    throw std::exception();
-  } 
-  return true;
-}
-
-map<int, string> parseOptions(int argc, char** argv) {
-  if (!meetsReq(argc, argv)) { throw std::exception(); }
-
-  fs::path imageFilePath;
-  fs::path audioFilePath;
-
-  for (int i = 0; i < argc; i++) {
-    string arg = string(argv[i]);
-
-    if (arg.compare("-h") || arg.compare("--help")) { showUsage(argv[0]); } 
-    if (arg.compare("-f") || arg.compare("--fast")) { bestQuality = true; }
-    if (Image::isImage(arg)) { imageFilePath = argv[i]; }
-    if (Audio::isAudio(arg)) { audioFilePath = argv[i]; }
-  } 
-
-  if (imageFilePath.empty() || audioFilePath.empty()) { throw std::exception(); }
-
-  map<int, string> result = {{0, imageFilePath}, {1, audioFilePath}};
-  return result; 
-}
-
 void cleanTempFiles(fs::path tempLogFile, fs::path tempAudioFile) {
   if (!tempLogFile.empty())   { remove(tempLogFile); }
   if (!tempAudioFile.empty()) { remove(tempAudioFile); }
@@ -191,16 +154,9 @@ void encodeImage(fs::path imageFilePath, fs::path encodedAudioFilePath, string e
   audioFileData.close();
 }
 
-int embed(int argc, char** argv) {
-  map<int, string> mediaFiles;
-  try {
-    mediaFiles = parseOptions(argc, argv);
-  } catch (const std::exception&) {
-    return -1;
-  }
-
-  fs::path audioFilePath = mediaFiles[0];
-  fs::path imageFilePath = mediaFiles[1];
+//int embed(int argc, char** argv) {
+int embed(fs::path audioFilePath, fs::path imageFilePath, bool quality) {
+  bestQuality = quality;
 
   ifstream imageFile(imageFilePath, ifstream::in | ifstream::binary);
   ofstream audioFile(audioFilePath, ifstream::out | ifstream::binary);
