+++ test/audio/CMakeLists.txt
@@ -3,7 +3,7 @@ add_executable(${TARGET} EmbedAudioTests.cpp)
 
 target_compile_features(${TARGET} PRIVATE cxx_std_20)
 target_link_libraries(${TARGET} PRIVATE audio_library image_library file_library fmt doctest)
-target_include_directories(${TARGET} PUBLIC ${DOCTEST_LIB} )
+target_include_directories(${TARGET} PUBLIC ${DOCTEST_LIB} ${HEADER_LIST})
 #FILE(COPY "../resources/inputFile1.png" DESTINATION "${CMAKE_BINARY_DIR}/test/audio")
 FILE(COPY "../resources/inputFile1.png" DESTINATION "${CMAKE_BINARY_DIR}")
