+++ test/audio/EmbedAudioTests.cpp
@@ -13,24 +13,23 @@
 using namespace std;
 namespace fs = std::filesystem;
 
-template <typename Map>
+//template <typename Map>
 
-bool key_compare (Map const &lhs, Map const &rhs) {
-    return lhs.size() == rhs.size()
-        && std::equal(lhs.begin(), lhs.end(), rhs.begin(), 
-                      [] (auto a, auto b) { return a.first == b.first; });
-}
+//bool key_compare (Map const &lhs, Map const &rhs) {
+    //return lhs.size() == rhs.size()
+        //&& std::equal(lhs.begin(), lhs.end(), rhs.begin(), 
+                      //[] (auto a, auto b) { return a.first == b.first; });
+//}
 
-map<int, string> createArgvInput(vector<string> arguments) {
-    vector<string> firstTestArgs = {""};
-    vector<char*> argv;
-    for (const auto& arg : arguments)
-        argv.push_back((char*)arg.data());
-    argv.push_back(nullptr);
+//map<int, string> createArgvInput(vector<string> arguments) {
+    //vector<string> firstTestArgs = {""};
+    //vector<char*> argv;
+    //for (const auto& arg : arguments)
+        //argv.push_back((char*)arg.data()); argv.push_back(nullptr);
 
-    map<int, string> filepaths = parseOptions(argv.size() - 1, argv.data()); 
-    return filepaths;
-}
+    //map<int, string> filepaths = parseOptions(argv.size() - 1, argv.data()); 
+    //return filepaths;
+//}
 
 TEST_CASE("Audio files can be embedded into image files") {
   SUBCASE("Test toLowerCase()") {
@@ -42,17 +41,17 @@ TEST_CASE("Audio files can be embedded into image files") {
     CHECK(File::toLowerCase(".PNG") == ".png");
   }
 
-  SUBCASE("Testing parseOptions()") { 
-    INFO("Current output of parseOptions: ");
+  //SUBCASE("Testing parseOptions()") { 
+    //INFO("Current output of parseOptions: ");
 
-    vector<string> firstTestArgs = {""};
-    CHECK_THROWS_AS(createArgvInput(firstTestArgs), std::exception);
+    //vector<string> firstTestArgs = {""};
+    //CHECK_THROWS_AS(createArgvInput(firstTestArgs), std::exception);
 
-    vector<string> secondTestArgs = {"image.png", "audio.ogg"};
+    //vector<string> secondTestArgs = {"image.png", "audio.ogg"};
     
-    map<int, string> expectedPaths = { {0, "image.png"}, {1, "audio.ogg"}};
-    CHECK(key_compare(expectedPaths, createArgvInput(secondTestArgs)));
-  } 
+    //map<int, string> expectedPaths = { {0, "image.png"}, {1, "audio.ogg"}};
+    //CHECK(key_compare(expectedPaths, createArgvInput(secondTestArgs)));
+  //} 
 
   //SUBCASE("Running image file checks") {
     //std::filesystem::path filepath = "../../inputFile1.png";
